fix: resolved migration
