--- alembic/versions/2e6cde290cc7_added_region_table_and_org_id_column_in_.py
@@ -1,60 +1,0 @@
-"""added region table and org_id column in contact us
-
-Revision ID: 2e6cde290cc7
-Revises: f8aca8048e80
-Create Date: 2024-08-01 00:38:47.956350
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '2e6cde290cc7'
-down_revision: Union[str, None] = '70dab65f6844'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('ix_timezones_id', table_name='timezones')
-    op.drop_table('timezones')
-    op.drop_index('ix_languages_id', table_name='languages')
-    op.drop_table('languages')
-    op.add_column('contact_us', sa.Column('org_id', sa.String(), nullable=False))
-    op.create_foreign_key(None, 'contact_us', 'organizations', ['org_id'], ['id'], ondelete='CASCADE')
-    op.add_column('regions', sa.Column('region', sa.String(), nullable=False))
-    op.add_column('regions', sa.Column('language', sa.String(), nullable=True))
-    op.add_column('regions', sa.Column('timezone', sa.String(), nullable=True))
-    op.drop_column('regions', 'name')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('regions', sa.Column('name', sa.VARCHAR(), autoincrement=False, nullable=False))
-    op.drop_column('regions', 'timezone')
-    op.drop_column('regions', 'language')
-    op.drop_column('regions', 'region')
-    op.drop_constraint(None, 'contact_us', type_='foreignkey')
-    op.drop_column('contact_us', 'org_id')
-    op.create_table('languages',
-    sa.Column('name', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('id', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('created_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=True),
-    sa.Column('updated_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name='languages_pkey')
-    )
-    op.create_index('ix_languages_id', 'languages', ['id'], unique=False)
-    op.create_table('timezones',
-    sa.Column('name', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('id', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('created_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=True),
-    sa.Column('updated_at', postgresql.TIMESTAMP(timezone=True), server_default=sa.text('now()'), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name='timezones_pkey')
-    )
-    op.create_index('ix_timezones_id', 'timezones', ['id'], unique=False)
-    # ### end Alembic commands ###
